refactor(teams): single file for project creation slice
